feat: add new username(s) to postgres database, and display them on homepage (#1)
* chore: install pg (short for node-postgres) and create pool

a pool is a pool of clients, which holds onto connections and opens a new connection unless there's an existing spare one, which is perfect for web servers. by contrast, a client is an _individual_ connection to the database, which is manually managed. you open a connection, do query, and close it. this is fine for once-off queries, but can become expensive if dealing with lots of queries -> need to hold onto client, which pool solves!

* feat(queries): add queries getAllUsernames and insertUsername

* feat: display list of usernames in database at / route, and insert new user name into db.

this included refactoring indexController to use db queries set up in previous commit

* chore: add comments for clarity

explains alt way of connection db pool using Connection URI, and fix typo in comment